requirements.txt
Django==4.1.2
psycopg2-binary
openpyxl
drf-yasg
djangorestframework
setuptools